Inital Start
Added form field for submission of image and background handling of file within UploadController.cs

UploadController is essentially inheriting Controller and then when the submit button is clicked the respective method within the UploadController class is called. The file is then evaluated and a new page is returned with the file info.